+++ src/vst3/mod.rs
@@ -6,6 +6,7 @@ pub mod component_handler;
 pub mod event_list;
 pub mod host_alloc;
 pub mod host_context;
+pub mod ibstream;
 pub mod instance;
 pub mod module;
 pub mod param_changes;